Shopping List App
This is my first flutter application. It's a shopping list application. We write the things we want to buy to the text field above and click add button on the right. The product is added to the list. If we want to make changes in the list we can do it with just one click while we want to delete a product from the list, we have to press and hold the screen.